+++ src/Propulsion.CosmosStore/CosmosStoreParser.fs
@@ -1,20 +1,13 @@
-#if COSMOSV2
-namespace Propulsion.Cosmos
-
-open Equinox.Cosmos.Store
-open Microsoft.Azure.Documents
-#else
 namespace Propulsion.CosmosStore
 
 open Equinox.CosmosStore.Core
-#endif
 
 open Propulsion.Streams
 
 /// Maps fields in an Event within an Equinox.Cosmos V1+ Event (in a Batch or Tip) to the interface defined by Propulsion.Streams
 /// <remarks>NOTE No attempt is made to filter out Tip (`id=-1`) batches from the ChangeFeed; Equinox versions >= 3, Tip batches can bear events.</remarks>
 [<RequireQualifiedAccess>]
-#if !COSMOSV2 && !COSMOSV3
+#if !COSMOSV3
 module EquinoxSystemTextJsonParser =
 
     type System.Text.Json.JsonDocument with
@@ -50,22 +43,6 @@ module EquinoxSystemTextJsonParser =
     let enumStreamEvents categoryFilter d : Default.StreamEvent seq =
         tryParseEquinoxBatch categoryFilter d |> ValueOption.map enumEquinoxCosmosEvents |> ValueOption.defaultValue Seq.empty
 #else
-#if COSMOSV2
-module EquinoxCosmosParser =
-
-    type Document with
-        member document.Cast<'T>() =
-            let tmp = Document()
-            tmp.SetPropertyValue("content", document)
-            tmp.GetPropertyValue<'T>("content")
-    type Batch with
-        member _.MapData x = x
-
-    /// Sanity check to determine whether the Document represents an `Equinox.Cosmos` >= 1.0 based batch
-    let isEquinoxBatch (d : Document) =
-        d.GetPropertyValue "p" <> null && d.GetPropertyValue "i" <> null
-        && d.GetPropertyValue "n" <> null && d.GetPropertyValue "e" <> null
-#else
 module EquinoxNewtonsoftParser =
 
     type Newtonsoft.Json.Linq.JObject with
@@ -81,7 +58,6 @@ module EquinoxNewtonsoftParser =
     /// Sanity check to determine whether the Document represents an `Equinox.Cosmos` >= 1.0 based batch
     let isEquinoxBatch (d : Newtonsoft.Json.Linq.JObject) =
         d.ContainsKey "p" && d.ContainsKey "i" && d.ContainsKey "n" && d.ContainsKey "e"
-#endif
 
     /// Enumerates the events represented within a batch
     let enumEquinoxCosmosEvents (batch : Batch) : Default.StreamEvent seq =